rename index to mapbox
