+++ packages/react-ui-map/src/index.js
@@ -1,5 +1,5 @@
-export { default } from './Mapbox'
-export { default as Mapbox } from './Mapbox'
+export { default } from './Mapbox/Mapbox'
+export { default as Mapbox } from './Mapbox/Mapbox'
 export { default as LayerInteractions } from './Mapbox/LayerInteractions'
 export { default as Source } from './Mapbox/Source'
 export { default as Layer } from './Mapbox/Layer'
